Get rid of circular import. Fix bug for suffix check when a list of images is provided.
